add one test for chip breakdown
