+++ package.json
@@ -17,6 +17,7 @@
     "react-dom": "^18.2.0",
     "react-image-crop": "^10.0.9",
     "react-scripts": "5.0.1",
+    "react-test": "^0.21.1",
     "typescript": "^4.4.4",
     "web-vitals": "^2.1.4"
   },
